fix NSInteger-long dataType bugs

+++ mgspoofhelper/MGAppPickerController.m
@@ -146,7 +146,7 @@ CFPropertyListRef MGCopyAnswer(CFStringRef);
 
 
 -(UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
-	NSString *cellIdentifier = [NSString stringWithFormat:@"AppPickerCellC%ldR%ld", indexPath.section, indexPath.row];
+	NSString *cellIdentifier = [NSString stringWithFormat:@"AppPickerCellC%ldR%ld", (long)indexPath.section, (long)indexPath.row];
 	UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:cellIdentifier];
 
 	if (!cell) {
